+++ feature/smart-config/src/main/java/com/buzbuz/smartautoclicker/feature/smart/config/ui/action/changecounter/ChangeCounterDialog.kt
@@ -122,8 +122,8 @@ class ChangeCounterDialog(
                     viewModel.setOperationValue(
                         if (checkedId == 0) {
                             CounterOperationValue.Number(
-                                if (editValueLayout.textField.text.isNullOrEmpty()) 0
-                                else editValueLayout.textField.text.toString().toInt()
+                                if (editValueLayout.textField.text.isNullOrEmpty()) 0.0
+                                else editValueLayout.textField.text.toString().toDouble()
                             )
                         } else {
                             CounterOperationValue.Counter(
@@ -141,8 +141,8 @@ class ChangeCounterDialog(
                 setOnTextChangedListener {
                     viewModel.setOperationValue(
                         CounterOperationValue.Number(
-                            if (editValueLayout.textField.text.isNullOrEmpty()) 0
-                            else editValueLayout.textField.text.toString().toInt()
+                            if (editValueLayout.textField.text.isNullOrEmpty()) 0.0
+                            else editValueLayout.textField.text.toString().toDouble()
                         )
                     )
                 }
