+++ feature/smart-debugging/src/main/java/com/buzbuz/smartautoclicker/feature/smart/debugging/ui/dialog/report/details/counter/DebugCounterStateContentUiState.kt
@@ -29,6 +29,6 @@ sealed class DebugCounterStateContentUiState {
 
 data class CounterStateItem(
     val counterName: String,
-    val currentCounterValue: Int,
-    val oldCounterValue: Int?,
+    val currentCounterValue: Double,
+    val oldCounterValue: Double?,
 )
